requirements.txt
semantic-kernel
ipykernel
azure-monitor-opentelemetry-exporter
semantic-kernel[azure]
fastapi
uvicorn
requests 
beautifulsoup4